remove test_context.py and append unit test to context.py
