+++ pyproject.toml
@@ -1,4 +1,4 @@
 [tool.pytest.ini_options]
-addopts = "-v -p no:warnings --doctest-modules --ignore=tests/test_context.py --show-progress -n 1 --dist=loadfile"
+addopts = "-v -p no:warnings --doctest-modules --show-progress -n 1 --dist=loadfile"
 python_files = 'test_*.py'
 testpaths = ['tests']
